Set up documentation building for docs.rs
If there is no readme per crate, then the crate documentation is
rendered, but this does not include the ability to hide code lines from
examples. This change adds stub readmes to the crates, and points the
documentation keys in the manifest files to the correct url on docs.rs.