Tweak documentation to be more consistent
Since we have methods on FitsHdu which really reflect trait methods
across various traits, it's good that we make the documentation more
consistent.

Previously the FitsHdu methods' documentation was very short and not
particularly useful. Instead the trait methods had more documentation.
However most users are likely to look up the methods on FitsHdu and find
it lacking.