Looser version specifier of `ndarray`
@Findus32 showed compilation problems with the homepage example. See
https://github.com/mindriot101/rust-fitsio/issues/102

The problem was caused by different versions of `ndarray`. `fitsio`
specified 0.11.2, and implemented `ReadImage` for the 0.11.2 version of
`ArrayD`, whereas the example did not specify the version of `ndarray`
(i.e. `ndarray = "*"`). When @Findus32 did a fresh clone of the repo, it
installed the latest version of `ndarray` i.e. 0.12.1.

It seems that implementing a trait for a different version of a struct
(i.e. `ArrayD` version 0.11.2 vs `ArrayD` version 0.12.1) does not
propagate to the newer version.

This might be quite annoying, if the version of `ndarray` is different
between library code and user code.