Implement fitsfile fptr via ptr::NonNull
This means the `fptr` pointer can never be null, which is a good
guarantee to have. On the other hand, we have to make all of the member
methods take `self` mutably (i.e. exclusive access).

I think this is probably a reasonable tradeoff as the fitsfile pointer
should be owned exclusively. Two concurrent alterations to the same
backed file should not happen.