Tweak details around the static-compile feature.
The feature was "static", now "fitsio-src".

Add a test to the CI bin/test script. I've also verified that
cfitsio-using code works correctly with and without the feature.

Make the behaviour of the bind_cfitsio function in fitsio-sys/build.rs
dependent on the feature.

Add a README to where the cfitsio source code lives, to help clarify the
situation.