Specify required patch versions for all deps.
Attempting to build/test after running `cargo +nightly -Z
minimal-versions update` will fail, because the dependency versions are
not tight enough. This commit specifies the minimum version of each
dependency that allows builds to succeed and tests to pass.

Note that if ndarray changes its API enough, using new versions of it
will cause fitsio to fail to compile; however, this was always the case
before this commit.