Re-export fitsio-sys as part of fitsio
I was getting conflicts between the versions of `fitsio-sys` that were
included in the crate compilation. Rust only allows one crate that links
with a system library to be included in the list of crates.

By exporting `fitsio-sys`, we can ensure that if the user needs
`fitsio-sys`, then it's the same version that's used by `fitsio` and
there will be no conflict.