Revert "Include release workflow"
This reverts commit 427d42520168312cdec87f114b3d991a1a3a803c.

The previous description was: this is not that useful. Now I will change
my description to be more useful...

This workflow only adds a commit bumping the version, but that won't
work with `fitsio` because we have tests ensuring that the readme points
to the latest version, which is not done with `cargo-release`.