fitsio-sys: switched CFITSIO build to use CMake
When cross-compiling to multiple architectures, fitsio-sys using the
Autotools path leads to build failures: the generated Makefile only
supports building CFITSIO into the source directory, i.e., a directory
within ~/.cargo most of the time. When separate builds need fitsio-sys
built for different architectures, this leads to file collisions and
(eventually) link errors.

Fortunately, CFITSIO has a CMake-based build path, and we switch to
using it in this commit. It did require a one-line fix in
CMakeLists.txt, which is included.